rename functions to integration
